LayoutTests/http/tests/security/contentSecurityPolicy/connect-src-eventsource-blocked-expected.txt

CONSOLE MESSAGE: Refused to connect to connect 'http://127.0.0.1:8000/eventsource/resources/simple-event-stream.asis' because it violates the following Content Security Policy directive: "connect-src http://localhost:8000".

Pass
